requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==44.0.3
Django==5.2.10
django-cors-headers==4.9.0
djangorestframework==3.16.1
pycparser==3.0
sqlparse==0.5.5
